requirements.txt
arrow==0.12.1
Django==2.0.5
django-constance==2.2.0
django-picklefield==1.0.0
djangorestframework==3.8.2
numpy==1.14.3
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
